requirements.txt
asgiref==3.2.9
attrs==19.3.0
autopep8==1.5.3
awsebcli==3.19.2
bcrypt==3.1.7
blessed==1.17.8
boto3
botocore
cached-property==1.5.1
cachetools==4.2.0
cement==2.8.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
Django==3.0.7
django-cors-headers==3.5.0
django-crontab==0.7.1
django-environ==0.4.5
django-filter==21.1
django-rest-passwordreset==1.1.0
django-storages==1.9.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
docker==4.2.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
google-api-core==1.25.0
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-translate==3.0.2
googleapis-common-protos==1.52.0
grpcio==1.34.1
idna==2.7
jmespath==0.10.0
jsonschema==3.2.0
libcst==0.3.16
mypy-extensions==0.4.3
numpy==1.19.0
pandas==1.0.5
paramiko==2.7.1
pathlib-mate==1.0.0
pathspec==0.5.9
Pillow==7.1.2
proto-plus==1.13.0
protobuf==3.14.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.44
requests==2.24.0
rsa==4.7
s3transfer==0.3.3
semantic-version==2.5.0
sentry-sdk==1.1.0
six==1.15.0
SQLAlchemy==1.3.18
sqlparse==0.3.1
termcolor==1.1.0
texttable==1.6.2
toml==0.10.1
twilio==6.2.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.11
uszipcode==0.2.4
wcwidth==0.1.9
websocket-client==0.57.0
xlrd==1.2.0
django-admin-rangefilter===0.8.4
XlsxWriter==3.0.3
